# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Character AI API
  description: "\n# Character AI API\n\nA comprehensive API for managing interactive\
    \ AI characters in toys and devices.\n\n## Features\n\n- **Character Management**:\
    \ Create, configure, and manage AI characters\n- **Real-time Interaction**: Process\
    \ audio input and generate character responses\n- **Session Memory**: Maintain\
    \ conversation context with configurable limits\n- **Safety & Security**: Content\
    \ filtering and safety analysis\n- **Voice Management**: Handle voice artifacts\
    \ and embeddings\n- **Authentication**: Device identity and JWT-based authentication\n\
    \n## Authentication\n\nThe API supports multiple authentication methods:\n\n1.\
    \ **Device Registration**: Register a new device and receive authentication tokens\n\
    2. **JWT Tokens**: Use Bearer tokens for authenticated requests\n3. **Admin Tokens**:\
    \ Special tokens for administrative operations\n\n## Rate Limiting\n\n- Default:\
    \ 60 requests per minute per device\n- Burst allowance: 10 requests\n- Rate limit\
    \ headers are included in responses\n\n## Safety & Privacy\n\n- All content is\
    \ analyzed for toxicity and PII\n- Child-safe content filtering\n- Configurable\
    \ safety thresholds\n- Privacy-first design with on-device processing\n\n## Character\
    \ Profiles\n\nCharacters are defined using YAML profiles with:\n- Personality\
...
    \        "
  version: 1.0.0
servers:
...
paths:
...
  /api/v1/personalization/preferences:
    post:
      tags:
      - Personalization
      summary: Update Preference
      description: Update a user preference.
      operationId: update_preference_api_v1_personalization_preferences_post
      parameters:
      - name: current_device
        in: query
        required: true
        schema:
          type: string
          description: Device ID
          title: Current Device
        description: Device ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreferenceUpdateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreferenceResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Personalization
      summary: Get User Preferences
      description: Get all user preferences.
      operationId: get_user_preferences_api_v1_personalization_preferences_get
      parameters:
      - name: current_device
        in: query
        required: true
        schema:
          type: string
          description: Device ID
          title: Current Device
        description: Device ID
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PreferenceResponse'
                title: Response Get User Preferences Api V1 Personalization Preferences
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/personalization/recommendations:
    get:
      tags:
      - Personalization
      summary: Get Character Recommendations
      description: Get personalized character recommendations.
      operationId: get_character_recommendations_api_v1_personalization_recommendations_get
      parameters:
      - name: max_recommendations
        in: query
        required: false
        schema:
          type: integer
          maximum: 20
          minimum: 1
          description: Maximum number of recommendations
          default: 5
          title: Max Recommendations
        description: Maximum number of recommendations
      - name: include_affinities
        in: query
        required: false
        schema:
          type: boolean
          description: Include affinity scores
          default: true
          title: Include Affinities
        description: Include affinity scores
      - name: current_device
        in: query
        required: true
        schema:
          type: string
          description: Device ID
          title: Current Device
        description: Device ID
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CharacterRecommendationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/personalization/conversation-style:
    post:
      tags:
      - Personalization
      summary: Get Adaptive Conversation Style
      description: Get adaptive conversation style for a character.
      operationId: get_adaptive_conversation_style_api_v1_personalization_conversation_style_post
      parameters:
      - name: current_device
        in: query
        required: true
        schema:
          type: string
          description: Device ID
          title: Current Device
        description: Device ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConversationStyleRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationStyleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/personalization/insights:
    get:
      tags:
      - Personalization
      summary: Get User Insights
      description: Get user personalization insights.
      operationId: get_user_insights_api_v1_personalization_insights_get
      parameters:
      - name: current_device
        in: query
        required: true
        schema:
          type: string
          description: Device ID
          title: Current Device
        description: Device ID
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInsightsResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/personalization/privacy:
    post:
      tags:
      - Personalization
      summary: Handle Privacy Request
      description: Handle privacy-related requests (export, clear, status).
      operationId: handle_privacy_request_api_v1_personalization_privacy_post
      parameters:
      - name: current_device
        in: query
        required: true
        schema:
          type: string
          description: Device ID
          title: Current Device
        description: Device ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrivacyRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrivacyResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/personalization/learn:
    post:
      tags:
      - Personalization
      summary: Learn From Conversation
      description: Learn from conversation data (for internal use).
      operationId: learn_from_conversation_api_v1_personalization_learn_post
      parameters:
      - name: current_device
        in: query
        required: true
        schema:
          type: string
          description: Device ID
          title: Current Device
        description: Device ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              title: Conversation Data
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                title: Response Learn From Conversation Api V1 Personalization Learn
                  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CharacterRecommendationResponse:
      properties:
        recommendations:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Recommendations
          description: List of recommended characters
        total_available:
          type: integer
          title: Total Available
          description: Total available characters
        user_affinities:
          additionalProperties:
            type: number
          type: object
          title: User Affinities
          description: User character affinities
      type: object
      required:
      - recommendations
      - total_available
      - user_affinities
      title: CharacterRecommendationResponse
      description: Response model for character recommendations.
...
    ConversationStyleRequest:
      properties:
        character_id:
          type: string
          title: Character Id
          description: Character ID for style adaptation
      type: object
      required:
      - character_id
      title: ConversationStyleRequest
      description: Request model for adaptive conversation style.
    ConversationStyleResponse:
      properties:
        character_id:
          type: string
          title: Character Id
          description: Character ID
        style_adaptations:
          additionalProperties: true
          type: object
          title: Style Adaptations
          description: Style adaptations
        user_preferences:
          additionalProperties: true
          type: object
          title: User Preferences
          description: Applied user preferences
      type: object
      required:
      - character_id
      - style_adaptations
      - user_preferences
      title: ConversationStyleResponse
      description: Response model for adaptive conversation style.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PreferenceResponse:
      properties:
        preference_type:
          type: string
          title: Preference Type
          description: Type of preference
        value:
          title: Value
          description: Preference value
        confidence:
          type: number
          title: Confidence
          description: Confidence score
        source:
          type: string
          title: Source
          description: Learning source
        last_updated:
          type: number
          title: Last Updated
          description: Last update timestamp
        interaction_count:
          type: integer
          title: Interaction Count
          description: Number of interactions
      type: object
      required:
      - preference_type
      - value
      - confidence
      - source
      - last_updated
      - interaction_count
      title: PreferenceResponse
      description: Response model for user preferences.
    PreferenceUpdateRequest:
      properties:
        preference_type:
          type: string
          title: Preference Type
          description: Type of preference to update
        value:
          title: Value
          description: Preference value
        confidence:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Confidence
          description: Confidence in the preference
          default: 0.8
        source:
          type: string
          title: Source
          description: Source of the preference
          default: explicit_feedback
      type: object
      required:
      - preference_type
      - value
      title: PreferenceUpdateRequest
      description: Request model for updating user preferences.
    PrivacyRequest:
      properties:
        action:
          type: string
          title: Action
          description: 'Privacy action: ''export'', ''clear'', or ''status'''
      type: object
      required:
      - action
      title: PrivacyRequest
      description: Request model for privacy operations.
    PrivacyResponse:
      properties:
        action:
          type: string
          title: Action
          description: Action performed
        success:
          type: boolean
          title: Success
          description: Success status
        data:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Data
          description: Exported data (if applicable)
        message:
          type: string
          title: Message
          description: Status message
      type: object
      required:
      - action
      - success
      - message
      title: PrivacyResponse
      description: Response model for privacy operations.
...
    UserInsightsResponse:
      properties:
        user_id:
          type: string
          title: User Id
          description: User ID
        profile_age_days:
          type: number
          title: Profile Age Days
          description: Age of profile in days
        total_preferences:
          type: integer
          title: Total Preferences
          description: Total number of preferences
        interaction_patterns:
          type: integer
          title: Interaction Patterns
          description: Number of interaction patterns
        character_affinities:
          additionalProperties:
            type: number
          type: object
          title: Character Affinities
          description: Character affinity scores
        preferences_summary:
          additionalProperties:
            additionalProperties: true
            type: object
          type: object
          title: Preferences Summary
          description: Summary of preferences
        learning_sources:
          additionalProperties:
            type: integer
          type: object
          title: Learning Sources
          description: Learning source counts
        privacy_compliance:
          type: boolean
          title: Privacy Compliance
          description: Privacy compliance status
      type: object
      required:
      - user_id
      - profile_age_days
      - total_preferences
      - interaction_patterns
      - character_affinities
      - preferences_summary
      - learning_sources
      - privacy_compliance
      title: UserInsightsResponse
      description: Response model for user insights.
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from device registration
    AdminToken:
      type: apiKey
      in: header
      name: x-admin-token
      description: Admin token for administrative operations
...
security:
- BearerAuth: []
- AdminToken: []
...